Separate COMP and CMSSW envs. Fixes #4300
Provides two new scripts:
  crab.sh: wrapper script that does scram unset env before running crab
commands. Also, if the command is a submit run the bootstrap script
  crab3bootstrap: executes all the operations that needs the CMSSW
environment saving the output to a tmp dir. Returns the tmp directory or
the error message in case of errors.

Backward compatibility is guaranteed since the bootstrapDone() function
is called to check if the clients needs to perform the CMSSW related
operations or the bootstrap script was run.

also refactored many parts of the client. We are providin now:
  flushMemoryLogger function that group the code used to save the output messages that
are stored in memory
  validateSubmitOptions to group the code that in the crab submit
command validate the command line options passed by the user
  setSubmitParserOptions to set the command line options of the client
submit command