submit --dryrun. print info about last trial
submit --dryrun needs to run for at least 25s in order not
to make its estimates based on very few small events.
If the first dryrun attempt is too short, the same cmsRun
comman is called again, with an increasing number of events.
with this commit, every trial reports how long it took and
how many events the next trial will process.